test(functions): type-check the test suite (343 errors -> 0)
`tsc -p tsconfig.test.json` reported 343 errors, all in functions/tests.
It now reports 0, and the package's `typecheck` script runs it after
`tsc --noEmit`, so `bun run typecheck` fails on a test type error in
functions (checked with a deliberately wrong test file: TS2322, exit 1).

Two public typing bugs in src, fixed without a runtime change:

- solveODE, freqz and zpk2tf were published as `any`. Their factories'
  dependencies are untyped, so the result inherited `any` from
  `typed(...)`. The create callbacks now declare `TypedFunction`, like the
  other 2,300 typed-function exports; results are `unknown`. This is what
  the untyped ODE callbacks (TS7006) exposed.
- The unit-valued physical constants (speedOfLight, planckConstant, ...)
  were declared as a local stub `{ fixPrefix: boolean }`, so
  `speedOfLight.toNumeric()` did not compile. The Unit dependency is
  core's Unit class, so they are now core's `UnitInstance`. The duplicate-
  symbol allowlist entry for the stub ("a narrow local view") is removed:
  the stub was also the constants' published type.

Test fixes, with no @ts-ignore/@ts-expect-error/@ts-nocheck or `any`, and
no assertion removed or weakened:

- Narrowed `unknown` typed-function results where they are used
  (`as Float64Array`, `as number[]`, `as UnitInstance`, local result
  interfaces such as `Spectrum` and `FFTResult`).
- `as Unit` became `as UnitInstance` (Unit is a value).
- Annotated ODE/event callback parameters; prefixed unused ones with `_`.
- tests/helpers/gpu-hardware.ts types the legacy `isFallbackAdapter` and
  `requestAdapterInfo` members with an optional-member interface; the
  feature detection is unchanged.
- dedup-bucketB: `buildAndSort` is generic over the factory type; the
  slice-2 file routes its deliberately mistyped inputs (nested numbers to
  filterRegExp, a 2-D array to reduce) through one commented helper, and
  reads the absent `initial` export through a Record view.
- Two test bugs the types caught: kmeans was passed `{ seed: 1 }`, an
  option it does not have (it seeds deterministically), so the argument
  is dropped; typed-bridge built a BigNumber with `new BigNumber(42)`,
  which calls the private constructor with the wrong arguments, and now
  uses `BigNumber.fromNumber(42)`. Both assertions are unchanged.

functions vitest: 287 files passed / 5 skipped, 4936 tests passed / 53
skipped, the same as before. typecheck 33/33; both consumer type checks
(Bundler, NodeNext) pass.

+++ functions/tests/typed-bitwise-wasm.test.ts
@@ -230,7 +230,7 @@ describe('typed bitwise — WASM fallback when module not loaded', () => {
     // which is the actual contract under test.
     const a = Int32Array.from([0xff00, 0x0ff0, 0x00ff, 0x1234]);
     const b = Int32Array.from([0xf0f0, 0x0f0f, 0xaaaa, 0x4321]);
-    const out = await bitAnd(a, b);
+    const out = (await bitAnd(a, b)) as Int32Array;
     expect(Array.from(out)).toEqual([
       0xff00 & 0xf0f0,
       0x0ff0 & 0x0f0f,
@@ -250,7 +250,7 @@ describe('typed bitwise — WASM fallback when module not loaded', () => {
       a[i] = i | 0;
       b[i] = (i ^ 0xdeadbeef) | 0;
     }
-    const out = await bitOr(a, b);
+    const out = (await bitOr(a, b)) as Int32Array;
     expect(out).toBeInstanceOf(Int32Array);
     expect(out.length).toBe(len);
     // Verify the first and last elements match the oracle.
